+++ esp-hal/CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Cpu::other()` is now marked as public (#4188)
 - The ESP_HAL_CONFIG_STACK_GUARD_MONITORING (enabled by default) enables a data watchpoint on the stack guard value to protect the main stack (#4207)
 - `start_app_core_with_stack_guard_offset` (#4207)
+- The `rmt::PulseCode::MAX_LEN` constant was added. (#XXXX)
 
 ### Changed